+++ user/src/main/java/com/logistcshub/user/application/service/DeliveryManagerService.java
@@ -6,8 +6,12 @@ import com.google.gson.reflect.TypeToken;
 import com.logistcshub.user.application.client.HubClient;
 import com.logistcshub.user.application.dtos.HubDto;
 import com.logistcshub.user.application.mapper.DeliveryManagerMapper;
-import com.logistcshub.user.application.security.UserDetailsImpl;
-import com.logistcshub.user.domain.model.*;
+import com.logistcshub.user.common.security.UserDetailsImpl;
+import com.logistcshub.user.domain.model.deliveryManager.DeliveryManager;
+import com.logistcshub.user.domain.model.deliveryManager.DeliveryManagerType;
+import com.logistcshub.user.domain.model.deliveryManager.HubManager;
+import com.logistcshub.user.domain.model.user.User;
+import com.logistcshub.user.domain.model.user.UserRoleEnum;
 import com.logistcshub.user.infrastructure.repository.DeliveryManagerRepository;
 import com.logistcshub.user.infrastructure.repository.HubManagerRepository;
 import com.logistcshub.user.infrastructure.repository.UserRepository;
