+++ user/src/main/java/com/logistcshub/user/application/service/HubManagerService.java
@@ -4,10 +4,10 @@ import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 import com.logistcshub.user.application.client.HubClient;
 import com.logistcshub.user.application.dtos.HubDto;
-import com.logistcshub.user.application.security.UserDetailsImpl;
-import com.logistcshub.user.domain.model.HubManager;
-import com.logistcshub.user.domain.model.User;
-import com.logistcshub.user.domain.model.UserRoleEnum;
+import com.logistcshub.user.common.security.UserDetailsImpl;
+import com.logistcshub.user.domain.model.deliveryManager.HubManager;
+import com.logistcshub.user.domain.model.user.User;
+import com.logistcshub.user.domain.model.user.UserRoleEnum;
 import com.logistcshub.user.infrastructure.repository.HubManagerRepository;
 import com.logistcshub.user.infrastructure.repository.UserRepository;
 import com.logistcshub.user.presentation.request.HubManagerRequest;
