+++ user/src/main/java/com/logistcshub/user/common/message/MessageType.java
@@ -1,4 +1,4 @@
-package com.logistcshub.user.infrastructure.common;
+package com.logistcshub.user.common.message;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
