+++ user/src/main/java/com/logistcshub/user/common/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.logistcshub.user.infrastructure.common;
+package com.logistcshub.user.common.response;
 
 import lombok.AccessLevel;
 import lombok.Builder;
