+++ user/src/main/java/com/logistcshub/user/domain/model/deliveryManager/DeliveryManager.java
@@ -1,6 +1,6 @@
-package com.logistcshub.user.domain.model;
+package com.logistcshub.user.domain.model.deliveryManager;
 
-import com.logistcshub.user.infrastructure.common.AuditEntity;
+import com.logistcshub.user.domain.model.AuditEntity;
 import jakarta.persistence.*;
 import lombok.*;
