+++ user/src/main/java/com/logistcshub/user/domain/model/deliveryManager/DeliveryManagerType.java
@@ -1,4 +1,4 @@
-package com.logistcshub.user.domain.model;
+package com.logistcshub.user.domain.model.deliveryManager;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
