+++ user/src/main/java/com/logistcshub/user/domain/model/deliveryManager/DeliveryStatus.java
@@ -1,4 +1,4 @@
-package com.logistcshub.user.domain.model;
+package com.logistcshub.user.domain.model.deliveryManager;
 
 public enum DeliveryStatus {
     ON_DELIVERY,
