+++ user/src/main/java/com/logistcshub/user/domain/repository/DeliveryManagerRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.logistcshub.user.domain.repository;
 
-import com.logistcshub.user.domain.model.UserRoleEnum;
+import com.logistcshub.user.domain.model.user.UserRoleEnum;
 import com.logistcshub.user.presentation.request.DeliSearchRequest;
 import com.logistcshub.user.presentation.response.DeliSearchResponse;
 import org.springframework.data.domain.Page;
