+++ user/src/main/java/com/logistcshub/user/infrastructure/repository/DeliveryManagerRepository.java
@@ -1,6 +1,6 @@
 package com.logistcshub.user.infrastructure.repository;
 
-import com.logistcshub.user.domain.model.DeliveryManager;
+import com.logistcshub.user.domain.model.deliveryManager.DeliveryManager;
 import com.logistcshub.user.domain.repository.DeliveryManagerRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
