+++ user/src/main/java/com/logistcshub/user/infrastructure/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.logistcshub.user.infrastructure.repository;
 
-import com.logistcshub.user.domain.model.User;
+import com.logistcshub.user.domain.model.user.User;
 import com.logistcshub.user.domain.repository.UserRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
