+++ user/src/main/java/com/logistcshub/user/presentation/controller/AuthController.java
@@ -3,8 +3,8 @@ package com.logistcshub.user.presentation.controller;
 import com.logistcshub.user.presentation.response.TokenDto;
 import com.logistcshub.user.presentation.response.UserDto;
 import com.logistcshub.user.application.service.AuthService;
-import com.logistcshub.user.infrastructure.common.ApiResponse;
-import com.logistcshub.user.infrastructure.common.MessageType;
+import com.logistcshub.user.common.response.ApiResponse;
+import com.logistcshub.user.common.message.MessageType;
 import com.logistcshub.user.presentation.request.LoginRequest;
 import com.logistcshub.user.presentation.request.SignupRequest;
 import jakarta.validation.Valid;
