+++ user/src/main/java/com/logistcshub/user/presentation/controller/UserController.java
@@ -3,11 +3,11 @@ package com.logistcshub.user.presentation.controller;
 import com.logistcshub.user.presentation.response.MyInfoDto;
 import com.logistcshub.user.presentation.response.UserSearchResponse;
 import com.logistcshub.user.presentation.response.UserDto;
-import com.logistcshub.user.application.security.UserDetailsImpl;
+import com.logistcshub.user.common.security.UserDetailsImpl;
 import com.logistcshub.user.application.service.UserService;
-import com.logistcshub.user.domain.model.UserRoleEnum;
-import com.logistcshub.user.infrastructure.common.ApiResponse;
-import com.logistcshub.user.infrastructure.common.MessageType;
+import com.logistcshub.user.domain.model.user.UserRoleEnum;
+import com.logistcshub.user.common.response.ApiResponse;
+import com.logistcshub.user.common.message.MessageType;
 import com.logistcshub.user.presentation.request.UserSearchRequest;
 import com.logistcshub.user.presentation.request.UserUpdateRequest;
 import lombok.RequiredArgsConstructor;
