+++ user/src/main/java/com/logistcshub/user/presentation/request/DeliSearchRequest.java
@@ -1,7 +1,7 @@
 package com.logistcshub.user.presentation.request;
 
-import com.logistcshub.user.domain.model.DeliveryManagerType;
-import com.logistcshub.user.domain.model.DeliveryStatus;
+import com.logistcshub.user.domain.model.deliveryManager.DeliveryManagerType;
+import com.logistcshub.user.domain.model.deliveryManager.DeliveryStatus;
 
 import java.io.Serializable;
 import java.util.UUID;
