+++ user/src/main/java/com/logistcshub/user/presentation/request/DeliveryManagerCreate.java
@@ -1,6 +1,6 @@
 package com.logistcshub.user.presentation.request;
 
-import com.logistcshub.user.domain.model.DeliveryManagerType;
+import com.logistcshub.user.domain.model.deliveryManager.DeliveryManagerType;
 
 import java.io.Serializable;
 import java.util.UUID;
