+++ user/src/main/java/com/logistcshub/user/presentation/request/UserSearchRequest.java
@@ -1,6 +1,6 @@
 package com.logistcshub.user.presentation.request;
 
-import com.logistcshub.user.domain.model.UserRoleEnum;
+import com.logistcshub.user.domain.model.user.UserRoleEnum;
 
 public record UserSearchRequest(Long userId, UserRoleEnum role, String username, String email) {
 }
